backend/requirements.txt
fastapi
uvicorn
pandas
numpy
plotly
google-generativeai
statsmodels
scikit-learn